refactor(extract): moves get-manifest
