+++ src/extract/resolve/get-manifest.mjs
