+++ src/extract/resolve/merge-manifests.mjs
