+++ test/extract/resolve/__fixtures__/empty-package-json/package.json
