+++ test/extract/resolve/__fixtures__/invalid-package-json/package.json
