+++ test/extract/resolve/__fixtures__/minimal-package-json/package.json
