+++ test/extract/resolve/__fixtures__/package-json-in-here/package.json
