+++ test/extract/resolve/__fixtures__/two-level-package-jsons/package.json
