+++ test/extract/resolve/__fixtures__/two-level-package-jsons/packages/subthing/package.json
